Nested Classes for XML reader
Se realizaron cambios para poder reducir el numero de archivos de clases
.
