+++ src/main/java/com/canfer/app/security/UserPrincipal.java
@@ -8,6 +8,7 @@ import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
+import com.canfer.app.model.Empresa;
 import com.canfer.app.model.Usuario;
 
 
@@ -68,6 +69,10 @@ public class UserPrincipal implements UserDetails {
 	
 	public Long getUserId() {
 		return this.usuario.getIdUsuario();
+	}
+	
+	public List<Empresa> getEmpresas() {
+		return this.usuario.getEmpresas();
 	}
 
 }
